+++ pkg/diff/compare_test.go
@@ -10,7 +10,7 @@ func TestDiffEmpty(t *testing.T) {
 	a := bitesize.Environment{}
 	b := bitesize.Environment{}
 
-	if d := Compare(a, b); d != "" {
+	if d := Compare(a, b); len(d) != 0 {
 		t.Errorf("Expected diff to be empty, got: %s", d)
 	}
 
@@ -22,7 +22,7 @@ func TestIgnoreTestFields(t *testing.T) {
 		{Name: "a"},
 	}}
 
-	if d := Compare(a, b); d != "" {
+	if d := Compare(a, b); len(d) != 0 {
 		t.Errorf("Expected diff to be empty, got: %s", d)
 	}
 }
@@ -33,7 +33,7 @@ func TestIgnoreDeploymentFields(t *testing.T) {
 		Method: "bluegreen",
 	}}
 
-	if d := Compare(a, b); d != "" {
+	if d := Compare(a, b); len(d) != 0 {
 		t.Errorf("Expected diff to be empty, got: %s", d)
 	}
 }
@@ -61,7 +61,7 @@ func TestIgnoreStatusFields(t *testing.T) {
 		},
 	}
 
-	if d := Compare(a, b); d != "" {
+	if d := Compare(a, b); len(d) != 0 {
 		t.Errorf("Expected diff to be empty, got: %s", d)
 	}
 }
@@ -70,7 +70,7 @@ func TestDiffNames(t *testing.T) {
 	a := bitesize.Environment{Name: "asd"}
 	b := bitesize.Environment{Name: "asdf"}
 
-	if Compare(a, b) != "" {
+	if len(Compare(a, b)) != 0 {
 		t.Error("Expected diff, got the same")
 	}
 }
@@ -79,12 +79,12 @@ func TestDiffVersionsSame(t *testing.T) {
 	var saTests = []struct {
 		versionA string
 		versionB string
-		expected bool
+		expected int
 	}{
-		{"1", "1", true},
-		{"1", "2", false},
-		{"", "1", true},  // assume the same if environments.bitesize does not have version
-		{"1", "", false}, // assume diff  if cluster is not deployed
+		{"1", "1", 0},
+		{"1", "2", 1},
+		{"", "1", 0}, // assume the same if environments.bitesize does not have version
+		{"1", "", 1}, // assume diff  if cluster is not deployed
 	}
 
 	for _, tst := range saTests {
@@ -95,7 +95,7 @@ func TestDiffVersionsSame(t *testing.T) {
 			Name: "a", Services: []bitesize.Service{{Name: "a", Version: tst.versionB}},
 		}
 
-		if res := Compare(a, b); (res == "") != tst.expected {
+		if res := Compare(a, b); len(res) != tst.expected {
 			t.Errorf(
 				"Unexpected version compare(%s,%s) should be %t\n%s\n A %+v\n B %+v",
 				tst.versionA, tst.versionB, tst.expected, res, a.Services, b.Services,